Refactor of k8stk to handle multiple subcommands
Replaced go-flags with cobra. Setup the main.go to just be a caller of the various
subcommands. Utilized the cobra generator which created the main.go and root.go.
This makes it so the root.go is called, which initializes all subcommands. Setup
merge as a subcommand. Converted the old main.go code to utilize cobra, and live
inside the new merge file.